Moved top of generator from Import to Generator
