+++ code/drasil-code/drasil-code.cabal
@@ -21,6 +21,7 @@ library
     , Language.Drasil.Code.Imperative.Doxygen.Import
     , Language.Drasil.Code.Imperative.Helpers
     , Language.Drasil.Code.Imperative.Data
+    , Language.Drasil.Code.Imperative.Generator
     , Language.Drasil.Code.Imperative.Import
     , Language.Drasil.Code.Imperative.LanguageRenderer
     , Language.Drasil.Code.Imperative.LanguageRenderer.JavaRenderer
@@ -72,6 +73,7 @@ executable codegenTest
     , Language.Drasil.Code.Imperative.Doxygen.Import
     , Language.Drasil.Code.Imperative.Helpers
     , Language.Drasil.Code.Imperative.Data
+    , Language.Drasil.Code.Imperative.Generator
     , Language.Drasil.Code.Imperative.Import
     , Language.Drasil.Code.Imperative.LanguageRenderer
     , Language.Drasil.Code.Imperative.LanguageRenderer.JavaRenderer
